Create an Anoma Escript that is not just Nockma
